# src/docs/swagger.yaml
openapi: 3.0.0
info:
    title: Scope manager API
    version: 1.0.0
    description: |
        REST API for managing organizations, their roles, element fields
        and agreement fields within Scope manager API.

        Each organization is the main container in the system: it groups elements,
        defines the vocabulary they must follow and establishes the available roles.
servers:
    - url: /
...
paths:
...
    /api/v1/organizations:
        get:
            tags:
                - Organizations
            summary: List all organizations
            operationId: getOrganizations
            security:
                - BearerAuth: []
            description: |
                Returns a paginated array with organizations registered in the system.
                Requires an authenticated user with system ADMIN role.
                Each organization includes its roles, element fields, agreement fields and members count.
            parameters:
                - $ref: '#/components/parameters/paginationPage'
                - $ref: '#/components/parameters/paginationLimit'
            responses:
                '200':
                    description: List of organizations retrieved successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OrganizationsResponse'
                            example:
                                success: true
                                message: OK
                                httpStatus: 200
                                appCode: SUCCESS
                                data:
                                    - _id: 665a1b2c3d4e5f6a7b8c9d0e
                                      name: universidad-sevilla
                                      displayName: Universidad de Sevilla
                                      description: Organization for Universidad de Sevilla
                                      createdBy: 665d4e5f6a7b8c9d0e1f2a3b
                                      elementFields: []
                                      agreementFields: []
                                      roles:
                                          - _id: 665b2c3d4e5f6a7b8c9d0e1f
                                            name: admin
                                            description: Organization Administrator
                                      members: 1
                                      createdAt: '2025-06-01T10:00:00.000Z'
                                      updatedAt: '2025-06-01T10:00:00.000Z'
                                      __v: 0
                                pagination:
                                    page: 1
                                    limit: 20
                                    totalItems: 1
                                    totalPages: 1
                                    hasNextPage: false
                                    hasPreviousPage: false
                                error: null
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '500':
                    $ref: '#/components/responses/InternalError'
        post:
            tags:
                - Organizations
            summary: Create an organization
            operationId: createOrganization
            security:
                - BearerAuth: []
            description: |
                Creates a new organization with a unique name, optional displayName and description.
                Requires an authenticated user with system ADMIN role.
                The organization is created with a default 'admin' role, stores the creator id
                and automatically assigns the creator as organization admin.
                The elementFields and agreementFields arrays are initialized empty.
            requestBody:
                $ref: '#/components/requestBodies/OrganizationBody'
            responses:
                '201':
                    description: Organization created successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OrganizationResponse'
                            example:
                                success: true
                                message: Organization created
                                httpStatus: 201
                                appCode: SUCCESS
                                data:
                                    _id: 665a1b2c3d4e5f6a7b8c9d0e
                                    name: universidad-sevilla
                                    displayName: Universidad de Sevilla
                                    description: Organization for Universidad de Sevilla
                                    createdBy: 665d4e5f6a7b8c9d0e1f2a3b
                                    elementFields: []
                                    agreementFields: []
                                    roles:
                                        - _id: 665b2c3d4e5f6a7b8c9d0e1f
                                          name: admin
                                          description: Organization Administrator
                                    createdAt: '2025-06-01T10:00:00.000Z'
                                    updatedAt: '2025-06-01T10:00:00.000Z'
                                    __v: 0
                                error: null
                '400':
                    $ref: '#/components/responses/ValidationError'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '409':
                    $ref: '#/components/responses/OrganizationConflict'
                '500':
                    $ref: '#/components/responses/InternalError'
    /api/v1/organizations/search:
        post:
            tags:
                - Organizations
            summary: Search organizations
            operationId: searchOrganizations
            security:
                - BearerAuth: []
            description: |
                Searches organizations using optional text filters and pagination.
                Requires an authenticated user.
                Filters can target name, displayName or both.
                Users with system ADMIN role search across all organizations.
                Non-admin users only search organizations where they are members.
                Each organization includes its members count.
            parameters:
                - $ref: '#/components/parameters/paginationPage'
                - $ref: '#/components/parameters/paginationLimit'
            requestBody:
                $ref: '#/components/requestBodies/OrganizationSearchBody'
            responses:
                '200':
                    description: Organizations matching the search filters
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OrganizationsResponse'
                            example:
                                success: true
                                message: OK
                                httpStatus: 200
                                appCode: SUCCESS
                                data:
                                    - _id: 665a1b2c3d4e5f6a7b8c9d0e
                                      name: universidad-sevilla
                                      displayName: Universidad de Sevilla
                                      description: Organization for Universidad de Sevilla
                                      createdBy: 665d4e5f6a7b8c9d0e1f2a3b
                                      elementFields: []
                                      agreementFields: []
                                      roles:
                                          - _id: 665b2c3d4e5f6a7b8c9d0e1f
                                            name: admin
                                            description: Organization Administrator
                                      members: 1
                                      createdAt: '2025-06-01T10:00:00.000Z'
                                      updatedAt: '2025-06-01T10:00:00.000Z'
                                      __v: 0
                                pagination:
                                    page: 1
                                    limit: 20
                                    totalItems: 1
                                    totalPages: 1
                                    hasNextPage: false
                                    hasPreviousPage: false
                                error: null
                '400':
                    $ref: '#/components/responses/ValidationError'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '500':
                    $ref: '#/components/responses/InternalError'
    /api/v1/organizations/{orgName}:
        parameters:
            - $ref: '#/components/parameters/orgName'
        get:
            tags:
                - Organizations
            summary: Get an organization by name
            operationId: getOrganizationByName
            security:
                - BearerAuth: []
            description: |
                Returns the full organization identified by its unique name,
                including its roles, element fields, agreement fields and members count.
                Requires an authenticated user that is either a member of the organization
                or has system ADMIN role.
            responses:
                '200':
                    description: Organization found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OrganizationResponse'
                            example:
                                success: true
                                message: OK
                                httpStatus: 200
                                appCode: SUCCESS
                                data:
                                    _id: 665a1b2c3d4e5f6a7b8c9d0e
                                    name: universidad-sevilla
                                    displayName: Universidad de Sevilla
                                    description: Organization for Universidad de Sevilla
                                    createdBy: 665d4e5f6a7b8c9d0e1f2a3b
                                    elementFields: []
                                    agreementFields: []
                                    roles:
                                        - _id: 665b2c3d4e5f6a7b8c9d0e1f
                                          name: admin
                                          description: Organization Administrator
                                    members: 1
                                    createdAt: '2025-06-01T10:00:00.000Z'
                                    updatedAt: '2025-06-01T10:00:00.000Z'
                                    __v: 0
                                error: null
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '404':
                    $ref: '#/components/responses/OrganizationNotFound'
                '500':
                    $ref: '#/components/responses/InternalError'
        put:
            tags:
                - Organizations
            summary: Update an organization
            operationId: updateOrganization
            security:
                - BearerAuth: []
            description: |
                Updates the name, displayName and description of an existing organization.
                Requires organization admin role or system ADMIN role.
                If the name is changed, the new name must be unique in the system.
                The request body is validated with the same required fields as creation.
            requestBody:
                $ref: '#/components/requestBodies/OrganizationBody'
            responses:
                '200':
                    description: Organization updated successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OrganizationResponse'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '409':
                    $ref: '#/components/responses/OrganizationConflict'
                '500':
                    $ref: '#/components/responses/InternalError'
        delete:
            tags:
                - Organizations
            summary: Delete an organization
            operationId: deleteOrganization
            security:
                - BearerAuth: []
            description: |
                Permanently deletes the organization and all its associated data.
                Requires organization admin role or system ADMIN role.
                Memberships and elements belonging to the organization are removed before the organization.
            responses:
                '200':
                    description: Organization deleted successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/NullResponse'
                            example:
                                success: true
                                message: Organization deleted
                                httpStatus: 200
                                appCode: SUCCESS
                                data: null
                                error: null
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'

    # ──────────────────────────────────────────────
    # Roles
    # ──────────────────────────────────────────────
    /api/v1/organizations/{orgName}/roles:
        parameters:
            - $ref: '#/components/parameters/orgName'
        post:
            tags:
                - Organizations Roles
            summary: Add a role to the organization
            operationId: addRole
            security:
                - BearerAuth: []
            description: |
                Adds a new role to the specified organization.
                Requires organization admin role or system ADMIN role.
                The role body must include name and description.
                The role name must be unique within the organization.
                There is a maximum role limit per organization (default 100).
            requestBody:
                $ref: '#/components/requestBodies/RoleBody'
            responses:
                '201':
                    description: Role added successfully. Returns the updated organization.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OrganizationResponse'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '409':
                    $ref: '#/components/responses/RoleConflict'
                '500':
                    $ref: '#/components/responses/InternalError'
    /api/v1/organizations/{orgName}/roles/{roleName}:
        parameters:
            - $ref: '#/components/parameters/orgName'
            - $ref: '#/components/parameters/roleName'
        put:
            tags:
                - Organizations Roles
            summary: Update a role
            operationId: updateRole
            security:
                - BearerAuth: []
            description: |
                Updates the name and/or description of an existing role.
                Requires organization admin role or system ADMIN role.
                The built-in 'admin' role cannot be modified.
                If the name is changed, the new name must be unique within the organization.
                The role body must include name and description.
            requestBody:
                $ref: '#/components/requestBodies/RoleBody'
            responses:
                '200':
                    description: Role updated successfully. Returns the updated organization.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OrganizationResponse'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/RoleNotFound'
                '409':
                    $ref: '#/components/responses/RoleConflict'
                '500':
                    $ref: '#/components/responses/InternalError'
        delete:
            tags:
                - Organizations Roles
            summary: Delete a role
            operationId: deleteRole
            security:
                - BearerAuth: []
            description: |
                Removes a role from the organization.
                Requires organization admin role or system ADMIN role.
                The built-in 'admin' role cannot be deleted.
                This operation also removes the role from associated memberships and deletes
                memberships that are left without roles.
            responses:
                '200':
                    description: Role deleted successfully. Returns the updated organization.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OrganizationResponse'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/RoleNotFound'
                '500':
                    $ref: '#/components/responses/InternalError'

    # ──────────────────────────────────────────────
    # Element Fields
    # ──────────────────────────────────────────────
    /api/v1/organizations/{orgName}/elementFields:
        parameters:
            - $ref: '#/components/parameters/orgName'
        post:
            tags:
                - Organizations ElementFields
            summary: Add an element field
            operationId: addElementField
            security:
                - BearerAuth: []
            description: |
                Adds a new field to the organization's element definition.
                Requires organization admin role or system ADMIN role.
                The field body must include name, description and type.
                Type must be one of `string`, `enum` or `number`.
                If type is `enum`, value must be an array; otherwise value must not be sent.
                This field becomes part of the structure that elements belonging
                to this organization must follow.
            requestBody:
                $ref: '#/components/requestBodies/FieldBody'
            responses:
                '201':
                    description: Element field added. Returns the updated organization.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OrganizationResponse'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/OrganizationNotFound'
                '409':
                    $ref: '#/components/responses/ElementFieldConflict'
                '500':
                    $ref: '#/components/responses/InternalError'
    /api/v1/organizations/{orgName}/elementFields/{fieldName}:
        parameters:
            - $ref: '#/components/parameters/orgName'
            - $ref: '#/components/parameters/elementFieldName'
        put:
            tags:
                - Organizations ElementFields
            summary: Update an element field
            operationId: updateElementField
            security:
                - BearerAuth: []
            description: |
                Updates an existing element field.
                Requires organization admin role or system ADMIN role.
                The existing field must belong to the organization's elementFields.
                The field body must include name, description and type.
                Type must be one of `string`, `enum` or `number`.
                If type is `enum`, value must be an array; otherwise value must not be sent.
                If the name is changed, the new name must be unique within the
                organization's elementFields.
            requestBody:
                $ref: '#/components/requestBodies/FieldBody'
            responses:
                '200':
                    description: Element field updated. Returns the updated organization.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OrganizationResponse'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/ElementFieldNotFound'
                '409':
                    $ref: '#/components/responses/ElementFieldConflict'
                '500':
                    $ref: '#/components/responses/InternalError'
        delete:
            tags:
                - Organizations ElementFields
            summary: Delete an element field
            operationId: deleteElementField
            security:
                - BearerAuth: []
            description: |
                Removes a field from the organization's element definition.
                Requires organization admin role or system ADMIN role.
                The field must exist in the organization's elementFields.
            responses:
                '200':
                    description: Element field deleted. Returns the updated organization.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OrganizationResponse'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/ElementFieldNotFound'
                '500':
                    $ref: '#/components/responses/InternalError'

    # ──────────────────────────────────────────────
    # Agreement Fields
    # ──────────────────────────────────────────────
...
components:
...
    securitySchemes:
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: JWT bearer token

    # ── Parameters ────────────────────────────────
    parameters:
...
        paginationPage:
            name: page
            in: query
            required: false
            description: Page number. Defaults to 1.
            schema:
                type: integer
                minimum: 1
                default: 1
            example: 1
        paginationLimit:
            name: limit
            in: query
            required: false
            description: Number of results per page. Defaults to 20.
            schema:
                type: integer
                minimum: 1
                default: 20
            example: 20
...
    responses:
...
        Unauthorized:
            description: Missing or invalid authentication token
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Authorization header missing or malformed
                        httpStatus: 401
                        appCode: UNAUTHORIZED
                        data: null
                        error:
                            message: Authorization header missing or malformed
                            details: null
...
        InternalError:
            description: Internal server error
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Internal server error
                        httpStatus: 500
                        appCode: UNKNOWN_ERROR
                        data: null
                        error:
                            message: Internal server error
                            details: null

    # ── Schemas ───────────────────────────────────
    schemas:
...
        OrganizationResponse:
            allOf:
                - $ref: '#/components/schemas/SuccessResponse'
                - type: object
                  properties:
                      data:
                          $ref: '#/components/schemas/Organization'
        OrganizationsResponse:
            allOf:
                - $ref: '#/components/schemas/SuccessResponse'
                - type: object
                  properties:
                      data:
                          type: array
                          items:
                              $ref: '#/components/schemas/Organization'
                      pagination:
                          $ref: '#/components/schemas/Pagination'
...